change_save_meta_in_zcc_crossponding_to_paddle and fix args
